+++ finetrainers/models/hunyuan_video/base_specification.py
@@ -12,6 +12,8 @@ from diffusers import (
 from diffusers.models.autoencoders.vae import DiagonalGaussianDistribution
 from transformers import AutoTokenizer, CLIPTextModel, CLIPTokenizer, LlamaModel
 
+from finetrainers.models.hunyuan_video import hunyuan_common
+
 from ... import data
 from ... import functional as FF
 from ...logging import get_logger
@@ -119,65 +121,11 @@ class HunyuanVideoModelSpecification(ModelSpecification):
     def _resolution_dim_keys(self):
         return {"latents": (2, 3, 4)}
 
-    def load_condition_models(self) -> Dict[str, torch.nn.Module]:
-        common_kwargs = {"revision": self.revision, "cache_dir": self.cache_dir}
+    load_condition_models = hunyuan_common.load_condition_models
 
-        if self.tokenizer_id is not None:
-            tokenizer = AutoTokenizer.from_pretrained(self.tokenizer_id, **common_kwargs)
-        else:
-            tokenizer = AutoTokenizer.from_pretrained(
-                self.pretrained_model_name_or_path, subfolder="tokenizer", **common_kwargs
-            )
+    load_latent_models = hunyuan_common.load_latent_models
 
-        if self.tokenizer_2_id is not None:
-            tokenizer_2 = AutoTokenizer.from_pretrained(self.tokenizer_2_id, **common_kwargs)
-        else:
-            tokenizer_2 = CLIPTokenizer.from_pretrained(
-                self.pretrained_model_name_or_path, subfolder="tokenizer_2", **common_kwargs
-            )
-
-        if self.text_encoder_id is not None:
-            text_encoder = LlamaModel.from_pretrained(
-                self.text_encoder_id, torch_dtype=self.text_encoder_dtype, **common_kwargs
-            )
-        else:
-            text_encoder = LlamaModel.from_pretrained(
-                self.pretrained_model_name_or_path,
-                subfolder="text_encoder",
-                torch_dtype=self.text_encoder_dtype,
-                **common_kwargs,
-            )
-
-        if self.text_encoder_2_id is not None:
-            text_encoder_2 = CLIPTextModel.from_pretrained(
-                self.text_encoder_2_id, torch_dtype=self.text_encoder_2_dtype, **common_kwargs
-            )
-        else:
-            text_encoder_2 = CLIPTextModel.from_pretrained(
-                self.pretrained_model_name_or_path,
-                subfolder="text_encoder_2",
-                torch_dtype=self.text_encoder_2_dtype,
-                **common_kwargs,
-            )
-
-        return {
-            "tokenizer": tokenizer,
-            "tokenizer_2": tokenizer_2,
-            "text_encoder": text_encoder,
-            "text_encoder_2": text_encoder_2,
-        }
-
-    def load_latent_models(self) -> Dict[str, torch.nn.Module]:
-        common_kwargs = {"revision": self.revision, "cache_dir": self.cache_dir}
-
-        if self.vae_id is not None:
-            vae = AutoencoderKLHunyuanVideo.from_pretrained(self.vae_id, torch_dtype=self.vae_dtype, **common_kwargs)
-        else:
-            vae = AutoencoderKLHunyuanVideo.from_pretrained(
-                self.pretrained_model_name_or_path, subfolder="vae", torch_dtype=self.vae_dtype, **common_kwargs
-            )
-
-        return {"vae": vae}
+    load_pipeline = hunyuan_common.load_pipeline
 
     def load_diffusion_models(self) -> Dict[str, torch.nn.Module]:
         common_kwargs = {"revision": self.revision, "cache_dir": self.cache_dir}
@@ -197,46 +145,6 @@ class HunyuanVideoModelSpecification(ModelSpecification):
         scheduler = FlowMatchEulerDiscreteScheduler()
 
         return {"transformer": transformer, "scheduler": scheduler}
-
-    def load_pipeline(
-        self,
-        tokenizer: Optional[AutoTokenizer] = None,
-        tokenizer_2: Optional[CLIPTokenizer] = None,
-        text_encoder: Optional[LlamaModel] = None,
-        text_encoder_2: Optional[CLIPTextModel] = None,
-        transformer: Optional[HunyuanVideoTransformer3DModel] = None,
-        vae: Optional[AutoencoderKLHunyuanVideo] = None,
-        scheduler: Optional[FlowMatchEulerDiscreteScheduler] = None,
-        enable_slicing: bool = False,
-        enable_tiling: bool = False,
-        enable_model_cpu_offload: bool = False,
-        training: bool = False,
-        **kwargs,
-    ) -> HunyuanVideoPipeline:
-        components = {
-            "tokenizer": tokenizer,
-            "tokenizer_2": tokenizer_2,
-            "text_encoder": text_encoder,
-            "text_encoder_2": text_encoder_2,
-            "transformer": transformer,
-            "vae": vae,
-            "scheduler": scheduler,
-        }
-        components = get_non_null_items(components)
-
-        pipe = HunyuanVideoPipeline.from_pretrained(
-            self.pretrained_model_name_or_path, **components, revision=self.revision, cache_dir=self.cache_dir
-        )
-        pipe.text_encoder.to(self.text_encoder_dtype)
-        pipe.text_encoder_2.to(self.text_encoder_2_dtype)
-        pipe.vae.to(self.vae_dtype)
-
-        _enable_vae_memory_optimizations(pipe.vae, enable_slicing, enable_tiling)
-        if not training:
-            pipe.transformer.to(self.transformer_dtype)
-        if enable_model_cpu_offload:
-            pipe.enable_model_cpu_offload()
-        return pipe
 
     @torch.no_grad()
     def prepare_conditions(
